+++ packages/app/src/renderer/src/pages/Tools/pages/Recorder/components/addModal.vue
@@ -77,6 +77,23 @@
           </template>
           <n-switch v-model:value="config.liveStartNotification" />
         </n-form-item>
+
+        <n-form-item>
+          <template #label>
+            <Tip
+              :text="textInfo.common.recorderType.text"
+              :tip="textInfo.common.recorderType.tip"
+            ></Tip>
+          </template>
+          <n-select
+            v-model:value="config.recorderType"
+            :options="recorderTypeOptions"
+            :disabled="globalFieldsObj.recorderType"
+          />
+          <n-checkbox v-model:checked="globalFieldsObj.recorderType" class="global-checkbox"
+            >全局</n-checkbox
+          >
+        </n-form-item>
         <n-form-item
           v-if="
             config.providerId !== 'Bilibili' &&
@@ -96,12 +113,6 @@
           <n-checkbox v-model:checked="globalFieldsObj.quality" class="global-checkbox"
             >全局</n-checkbox
           >
-        </n-form-item>
-        <n-form-item v-if="config.providerId !== 'HuYa' && config.providerId !== 'DouYin'">
-          <template #label>
-            <Tip text="只录制音频" tip="会选择纯音频流，B站只支持flv流，抖音请在画质中选择"></Tip>
-          </template>
-          <n-switch v-model:value="config.onlyAudio" />
         </n-form-item>
 
         <template v-if="config.providerId === 'Bilibili'">
@@ -323,6 +334,12 @@
           </n-form-item>
         </template>
 
+        <n-form-item v-if="config.providerId !== 'HuYa' && config.providerId !== 'DouYin'">
+          <template #label>
+            <Tip text="只录制音频" tip="会选择纯音频流，B站只支持flv流，抖音请在画质中选择"></Tip>
+          </template>
+          <n-switch v-model:value="config.onlyAudio" />
+        </n-form-item>
         <n-form-item>
           <template #label>
             <Tip :text="textInfo.common.format.text" :tip="textInfo.common.format.tip"></Tip>
@@ -479,6 +496,7 @@ import {
   videoFormatOptions,
   douyinStreamFormatOptions,
   huyaSourceOptions,
+  recorderTypeOptions,
 } from "@renderer/enums/recorder";
 import { useConfirm } from "@renderer/hooks";
 
@@ -512,6 +530,7 @@ const globalFieldsObj = ref<Record<NonNullable<Recorder["noGlobalFollowFields"]>
     codecName: true,
     source: true,
     videoFormat: true,
+    recorderType: true,
     cookie: true,
     doubleScreen: true,
     useServerTimestamp: true,
@@ -541,6 +560,7 @@ const config = ref<Omit<Recorder, "id">>({
   liveStartNotification: false,
   source: "auto",
   videoFormat: "auto",
+  recorderType: "ffmpeg",
   cookie: "",
   doubleScreen: true,
   useServerTimestamp: true,
@@ -654,6 +674,7 @@ watch(showModal, async (val) => {
       liveStartNotification: false,
       source: "auto",
       videoFormat: "auto",
+      recorderType: "ffmpeg",
       cookie: "",
       doubleScreen: true,
       useServerTimestamp: true,
@@ -679,6 +700,7 @@ watch(showModal, async (val) => {
       codecName: !(config.value?.noGlobalFollowFields ?? []).includes("codecName"),
       source: !(config.value?.noGlobalFollowFields ?? []).includes("source"),
       videoFormat: !(config.value?.noGlobalFollowFields ?? []).includes("videoFormat"),
+      recorderType: !(config.value?.noGlobalFollowFields ?? []).includes("recorderType"),
       cookie: !(config.value?.noGlobalFollowFields ?? []).includes("cookie"),
       doubleScreen: !(config.value?.noGlobalFollowFields ?? []).includes("doubleScreen"),
       useServerTimestamp: !(config.value?.noGlobalFollowFields ?? []).includes(
@@ -753,6 +775,9 @@ watch(
     }
     if (val.videoFormat) {
       config.value.videoFormat = appConfig.value.recorder.videoFormat;
+    }
+    if (val.recorderType) {
+      config.value.recorderType = appConfig.value.recorder.recorderType;
     }
     if (val.cookie) {
       config.value.cookie = appConfig.value.recorder.douyin.cookie;