+++ packages/liveManager/src/recorder/FFMPEGRecorder.ts
@@ -1,7 +1,7 @@
 import EventEmitter from "node:events";
 
-import { createFFMPEGBuilder, StreamManager, utils } from "./index.js";
-import { createInvalidStreamChecker, assert } from "./utils.js";
+import { createFFMPEGBuilder, StreamManager, utils } from "../index.js";
+import { createInvalidStreamChecker, assert } from "../utils.js";
 
 export class FFMPEGRecorder extends EventEmitter {
   private command: ReturnType<typeof createFFMPEGBuilder>;
@@ -15,6 +15,8 @@ export class FFMPEGRecorder extends EventEmitter {
   isHls: boolean;
   disableDanma: boolean = false;
   url: string;
+  formatName: "flv" | "ts" | "fmp4";
+  videoFormat: "ts" | "mkv" | "mp4";
   headers:
     | {
         [key: string]: string | undefined;
@@ -28,9 +30,9 @@ export class FFMPEGRecorder extends EventEmitter {
       segment: number;
       outputOptions: string[];
       inputOptions?: string[];
-      isHls?: boolean;
       disableDanma?: boolean;
-      videoFormat?: "auto" | "ts" | "mkv";
+      videoFormat?: "auto" | "ts" | "mkv" | "mp4";
+      formatName?: "flv" | "ts" | "fmp4";
       headers?: {
         [key: string]: string | undefined;
       };
@@ -40,12 +42,40 @@ export class FFMPEGRecorder extends EventEmitter {
   ) {
     super();
     const hasSegment = !!opts.segment;
+    this.hasSegment = hasSegment;
+
+    let formatName: "flv" | "ts" | "fmp4" = "flv";
+    if (opts.url.includes(".m3u8")) {
+      formatName = "ts";
+    }
+    this.formatName = opts.formatName ?? formatName;
+
+    if (this.formatName === "fmp4" || this.formatName === "ts") {
+      this.isHls = true;
+    } else {
+      this.isHls = false;
+    }
+
+    let videoFormat = opts.videoFormat ?? "auto";
+    if (videoFormat === "auto") {
+      if (!this.hasSegment) {
+        videoFormat = "mp4";
+        if (this.formatName === "ts") {
+          videoFormat = "ts";
+        }
+      } else {
+        videoFormat = "ts";
+      }
+    }
+    this.videoFormat = videoFormat;
+
     this.disableDanma = opts.disableDanma ?? false;
     this.streamManager = new StreamManager(
       opts.getSavePath,
       hasSegment,
       this.disableDanma,
-      opts.videoFormat,
+      "ffmpeg",
+      this.videoFormat,
       {
         onUpdateLiveInfo: this.onUpdateLiveInfo,
       },
@@ -55,21 +85,14 @@ export class FFMPEGRecorder extends EventEmitter {
       3 * 10e3,
       false,
     );
-    this.hasSegment = hasSegment;
     this.getSavePath = opts.getSavePath;
     this.ffmpegOutputOptions = opts.outputOptions;
     this.inputOptions = opts.inputOptions ?? [];
     this.url = opts.url;
     this.segment = opts.segment;
     this.headers = opts.headers;
-    if (opts.isHls === undefined) {
-      this.isHls = this.url.includes("m3u8");
-    } else {
-      this.isHls = opts.isHls;
-    }
 
     this.command = this.createCommand();
-
     this.streamManager.on("videoFileCreated", ({ filename, cover }) => {
       this.emit("videoFileCreated", { filename, cover });
     });
@@ -110,16 +133,17 @@ export class FFMPEGRecorder extends EventEmitter {
       .on("end", () => this.onEnd("finished"))
       .on("stderr", async (stderrLine) => {
         assert(typeof stderrLine === "string");
-        await this.streamManager.handleVideoStarted(stderrLine);
         this.emit("DebugLog", { type: "ffmpeg", text: stderrLine });
 
+        const [isInvalid, reason] = isInvalidStream(stderrLine);
+        if (isInvalid) {
+          this.onEnd(reason);
+        }
+
+        await this.streamManager.handleVideoStarted(stderrLine);
         const info = this.formatLine(stderrLine);
         if (info) {
           this.emit("progress", info);
-        }
-
-        if (isInvalidStream(stderrLine)) {
-          this.onEnd("invalid stream");
         }
       })
       .on("stderr", this.timeoutChecker?.update);