+++ packages/shared/src/index.ts
@@ -11,6 +11,7 @@ import logger, { initLogger, setLogLevel } from "./utils/log.js";
 import { migrateBiliUser, checkAccountLoop } from "./task/bili.js";
 import BiliCheckQueue from "./task/BiliCheckQueue.js";
 import { createInterval as checkSubLoop } from "./task/videoSub.js";
+import { check as checkVirtualRecordLoop } from "./task/virtualRecord.js";
 import { createRecorderManager } from "./recorder/index.js";
 import { sendNotify } from "./notify.js";
 import { initDB } from "./db/index.js";
@@ -25,6 +26,7 @@ const init = async (config: GlobalConfig) => {
     ffmpegPath: config.defaultFfmpegPath,
     ffprobePath: config.defaultFfprobePath,
     danmuFactoryPath: config.defaultDanmakuFactoryPath,
+    mesioPath: config.defaultMesioPath,
   });
   const logLevel = appConfig.get("logLevel");
   initLogger(config.logPath, logLevel);
@@ -58,6 +60,7 @@ const init = async (config: GlobalConfig) => {
     checkAccountLoop();
     checkDiskSpaceLoop();
     checkSubLoop();
+    checkVirtualRecordLoop();
   } catch (error) {
     logger.error("初始化失败", error);
   }