+++ packages/DouYuRecorder/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.6.0
+
+- 支持 `recorderType` 参数用于配置底层录制器，支持`auto | ffmpeg | mesio`
+- 修复cdn错误显示
+- 增加更多礼物
+- 修复 `videoFormat=auto` 时某些情况下格式的判断
+
 # 1.5.1
 
 - `resolveChannelInfoFromURL` 新增返回参数：`avatar`